seperate --keep and --nomerge, rename --nomerge to --no-merge

+++ src/core/archive.ts
@@ -59,7 +59,7 @@ class ArchiveDownloader extends Downloader {
             format,
             cookies,
             headers,
-            nomerge,
+            noMerge,
             keepEncryptedChunks,
             chunkNamingStrategy,
             cliMode,
@@ -76,7 +76,7 @@ class ArchiveDownloader extends Downloader {
             format,
             cookies,
             headers,
-            nomerge,
+            noMerge,
             keepEncryptedChunks,
             chunkNamingStrategy,
             cliMode,
@@ -463,14 +463,16 @@ class ArchiveDownloader extends Downloader {
             }
             logger.info("Merging chunks...");
             await this.fileConcentrator.waitAllFilesWritten();
-            logger.info("End of merging.");
-            logger.info("Starting cleaning temporary files.");
-            try {
-                await deleteEmptyDirectory(this.tempPath);
-            } catch (e) {
-                logger.warning(
-                    `Fail to delete temporary files, please delete manually or execute "minyami --clean" later.`
-                );
+            if (!this.keepTemporaryFiles) {
+                logger.info("End of merging.");
+                logger.info("Starting cleaning temporary files.");
+                try {
+                    await deleteEmptyDirectory(this.tempPath);
+                } catch (e) {
+                    logger.warning(
+                        `Fail to delete temporary files, please delete manually or execute "minyami --clean" later.`
+                    );
+                }
             }
             try {
                 deleteTask(this.m3u8Path.split("?")[0]);
